+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- New totalizers field to the affiliateOrders query
+
 ## [0.18.0] - 2022-03-08
 
 ### Added